Update name prefix
The script created a name prefix for the output file based on the basename
of the corpus path. However, because this is a directory, no name prefix was
used. This is inconvenient. Solved the problem by adding an option --name-prefix
to specify the prefix used, or use the name of the directory the corpus is stored
in (just the directory name, not the entire path) (default).